Revamp block logic to track full blocks (#13)
bug: Takes full block into account when making indentation decisions
  -  No-longer just waits for and looks at the first indented line
bug: Allows block-content to be at the SAME level as the block start/stop statements
bug: Ignores empty lines in indentation tracking
  - ANY content, including just spaces/tabs, is considered non-empty
bug: Makes indentation of text lines in script blocks more predictable

chore: Caches and inspects entire blocks before making indentation-tracking decisions, then re-processes the blocks.
chore: Removes line number tracking in debugging.
- Will add again later in a new way
chore: Moves line processing to process_line() to allow lines to be processed multiple times.

Breaking Change: The previous release introduced logic to find the first indented line in a block, then use that for tracking, skipping un-indented lines (or lines even with the block stop/start).  In practice, this created undesirable results for template writings who use certain (and valid) indention techniques.  This release replaces that logic, so your rendered templates may have different whitespace than they did previously.

Breaking Change: Now that we inspect the entire block, and not just the first line, to make indentation-tracking decisions, it becomes extra important to have consistent use of whitespace and indentation in your templates.  If any line in the block is not indented, then indentation tracking is mostly disabled for that block and results are undefined.
NOTE: it is OK to have empty lines in your blocks as they are ignored, but **ANY** content on the line (even if its just spaces or tabs) is considered text.